+++ tests/integration/store/test_version_store.py
@@ -17,6 +17,7 @@ import arctic
 from arctic.exceptions import NoDataFoundException, DuplicateSnapshotException, ArcticException
 from arctic.date import DateRange
 from arctic.store import _version_store_utils
+from tests.integration.chunkstore.test_utils import create_test_data
 
 from ...util import read_str_as_pandas
 from arctic.date._mktz import mktz
@@ -1068,7 +1069,7 @@ def test_write_metadata_followed_by_append(library):
     with patch('arctic.arctic.logger.info') as info:
         library.write(symbol, data=mydf_a, metadata={'field_a': 1})  # creates version 1
         library.write_metadata(symbol, metadata={'field_b': 1})  # creates version 2 (only metadata)
-        library.append(symbol, data=mydf_b,metadata={'field_c': 1})  # creates version 3
+        library.append(symbol, data=mydf_b, metadata={'field_c': 1})  # creates version 3
 
         # Trigger GC now
         library._prune_previous_versions(symbol, 0)
@@ -1182,18 +1183,18 @@ def test_restore_version(library):
         library.write(symbol, data=mydf_a, metadata={'field_a': 1})  # creates version 1
         library.write(symbol, data=mydf_b, metadata={'field_a': 2})  # creates version 2
 
-        v = library.read(symbol)
-        assert_frame_equal(v.data, mydf_b)
-        assert v.metadata == {'field_a': 2}
+        item = library.read(symbol)
+        assert_frame_equal(item.data, mydf_b)
+        assert item.metadata == {'field_a': 2}
         assert library._read_metadata(symbol).get('version') == 2
 
-        library.restore_version(symbol, as_of=1)  # creates version 3
+        restore_item = library.restore_version(symbol, as_of=1)  # creates version 3
+        assert restore_item.version == 3
+        assert restore_item.metadata == {'field_a': 1}
 
-        #library._delete_version(symbol, 1)  # delete the original version to test further the robustness/dependency
-
-        v = library.read(symbol)
-        assert_frame_equal(v.data, mydf_a)
-        assert v.metadata == {'field_a': 1}
+        item = library.read(symbol)
+        assert_frame_equal(item.data, mydf_a)
+        assert item.metadata == {'field_a': 1}
         assert library._read_metadata(symbol).get('version') == 3
 
 
@@ -1205,16 +1206,20 @@ def test_restore_version_followed_by_append(library):
     with patch('arctic.arctic.logger.info') as info:
         library.write(symbol, data=mydf_a, metadata={'field_a': 1})  # creates version 1
         library.write(symbol, data=mydf_b, metadata={'field_b': 2})  # creates version 2
-        library.restore_version(symbol, as_of=1)  # creates version 3
+
+        restore_item = library.restore_version(symbol, as_of=1)  # creates version 3
+        assert restore_item.version == 3
+        assert restore_item.metadata == {'field_a': 1}
+
         library.append(symbol, data=mydf_c, metadata={'field_c': 3})  # creates version 4
 
         # Trigger GC now
         library._prune_previous_versions(symbol, 0)
         time.sleep(2)
 
-        v = library.read(symbol)
-        assert_frame_equal(v.data, mydf_a.append(mydf_c))
-        assert v.metadata == {'field_c': 3}
+        item = library.read(symbol)
+        assert_frame_equal(item.data, mydf_a.append(mydf_c))
+        assert item.metadata == {'field_c': 3}
         assert library._read_metadata(symbol).get('version') == 4
 
 
@@ -1226,7 +1231,9 @@ def test_restore_version_purging_previous_versions(library):
         library.write(symbol, data=mydf_a, metadata={'field_a': 1})  # creates version 1
         library.write(symbol, data=mydf_b, metadata={'field_a': 2})  # creates version 2
 
-        library.restore_version(symbol, as_of=1)  # creates version 3
+        restore_item = library.restore_version(symbol, as_of=1)  # creates version 3
+        assert restore_item.version == 3
+        assert restore_item.metadata == {'field_a': 1}
 
         # Trigger GC now
         library._prune_previous_versions(symbol, 0)
@@ -1234,9 +1241,9 @@ def test_restore_version_purging_previous_versions(library):
 
         # library._delete_version(symbol, 1)  # delete the original version to test further the robustness/dependency
 
-        v = library.read(symbol)
-        assert_frame_equal(v.data, mydf_a)
-        assert v.metadata == {'field_a': 1}
+        item = library.read(symbol)
+        assert_frame_equal(item.data, mydf_a)
+        assert item.metadata == {'field_a': 1}
         assert library._read_metadata(symbol).get('version') == 3
 
 
@@ -1249,10 +1256,10 @@ def test_restore_version_non_existent_version(library):
         with pytest.raises(NoDataFoundException):
             library.restore_version(symbol, as_of=3)
 
-        v = library.read(symbol)
-        assert_frame_equal(v.data, mydf_a)
-        assert v.metadata == {'field_a': 1}
-        assert library._read_metadata(symbol).get('version') == 1
+        item = library.read(symbol)
+        assert_frame_equal(item.data, mydf_a)
+        assert item.metadata == {'field_a': 1}
+        assert item.version == 1
 
 
 def test_restore_version_which_updated_only_metadata(library):
@@ -1264,29 +1271,103 @@ def test_restore_version_which_updated_only_metadata(library):
         library.write_metadata(symbol, metadata={'field_b': 1})  # creates version 2
         library.write(symbol, data=mydf_b)  # creates version 3
 
-        library.restore_version(symbol, as_of=2)  # creates version 4
+        restore_item = library.restore_version(symbol, as_of=2)  # creates version 4
+        assert restore_item.version == 4
+        assert restore_item.metadata == {'field_b': 1}
 
-        v = library.read(symbol)
-        assert_frame_equal(v.data, mydf_a)
-        assert v.metadata == {'field_b': 1}
-        assert library._read_metadata(symbol).get('version') == 4
+        item = library.read(symbol)
+        assert_frame_equal(item.data, mydf_a)
+        assert item.metadata == {'field_b': 1}
+        assert item.version == 4
 
 
-def test_restore_version_snapshot(library):
+def test_restore_version_then_snapshot(library):
     symbol = 'FTL'
     mydf_a = _rnd_df(10, 5)
     mydf_b = _rnd_df(10, 5)
     with patch('arctic.arctic.logger.info') as info:
         library.write(symbol, data=mydf_a, metadata={'field_a': 1})  # creates version 1
         library.write_metadata(symbol, metadata={'field_b': 1})  # creates version 2
-        library.restore_version(symbol, as_of=2)  # creates version 3
+
+        restore_item = library.restore_version(symbol, as_of=1)  # creates version 3
+        assert restore_item.metadata == {'field_a': 1}
+        assert restore_item.version == 3
+
         library.snapshot('SNAP_1')
         library.write(symbol, data=mydf_b)  # creates version 3
 
-        v = library.read(symbol, as_of='SNAP_1')
-        assert_frame_equal(v.data, mydf_a)
-        assert v.metadata == {'field_b': 1}
-        assert library._read_metadata(symbol, as_of='SNAP_1').get('version') == 3
+        item = library.read(symbol, as_of='SNAP_1')
+        assert_frame_equal(item.data, mydf_a)
+        assert item.metadata == {'field_a': 1}
+        assert item.version == 3
+
+
+def test_restore_version_latest_snapshot_noop(library):
+    symbol = 'FTL'
+    mydf_a = _rnd_df(10, 5)
+    with patch('arctic.arctic.logger.info') as info:
+        library.write(symbol, data=mydf_a, metadata={'field_a': 1})  # creates version 1
+        library.write_metadata(symbol, metadata={'field_b': 1})  # creates version 2
+        library.snapshot('SNAP_1')
+
+        restore_item = library.restore_version(symbol, as_of='SNAP_1')  # does not create a new version
+        assert restore_item.metadata == {'field_b': 1}
+        assert restore_item.version == 2
+
+        item = library.read(symbol)
+        assert_frame_equal(item.data, mydf_a)
+        assert item.metadata == {'field_b': 1}
+        assert item.version == 2
+
+
+def test_restore_version_latest_version_noop(library):
+    symbol = 'FTL'
+    mydf_a = _rnd_df(10, 5)
+    with patch('arctic.arctic.logger.info') as info:
+        library.write(symbol, data=mydf_a, metadata={'field_a': 1})  # creates version 1
+        library.write_metadata(symbol, metadata={'field_b': 1})  # creates version 2
+
+        restore_item = library.restore_version(symbol, as_of=2)  # does not create a new version
+        assert restore_item.metadata == {'field_b': 1}
+        assert restore_item.version == 2
+
+        item = library.read(symbol)
+        assert_frame_equal(item.data, mydf_a)
+        assert item.metadata == {'field_b': 1}
+        assert item.version == 2
+
+
+def test_restore_version_snap_delete_symbol_restore(library):
+    symbol = 'FTL'
+    mydf = _rnd_df(20, 5)
+    with patch('arctic.arctic.logger.info') as info:
+        library.write(symbol, data=mydf[:10], metadata={'field_a': 1})  # creates version 1
+        library.append(symbol, data=mydf[10:15])  # version 2
+        library.snapshot('snapA')
+
+        library.append(symbol, data=mydf[15:20])  # version 3
+        library.delete(symbol)  # version 4
+
+        restored_item = library.restore_version(symbol, as_of='snapA')  # version 5
+        assert restored_item.metadata == {'field_a': 1}
+        assert restored_item.version == 5
+
+        item = library.read(symbol)
+        assert_frame_equal(item.data, mydf[:15])
+        assert item.metadata == {'field_a': 1}
+        assert item.version == 5
+
+
+def test_restore_from_version_with_deleted_symbol(library):
+    symbol = 'FTL'
+    mydf_a = _rnd_df(10, 5)
+    with patch('arctic.arctic.logger.info') as info:
+        library.write(symbol, data=mydf_a, metadata={'field_a': 1})  # creates version 1
+        library.delete(symbol)
+
+        with pytest.raises(NoDataFoundException):
+            library.restore_version(symbol, as_of=2)
+
 
 
 def test_prune_previous_versions_retries_on_cleanup_error(library):
@@ -1401,3 +1482,252 @@ def test_empty_string_column_name(library):
 
     with pytest.raises(ArcticException):
         library.write('df', df)
+
+
+def n_append(library, library_name, total_appends, rows_per_append, bulk_data_ts, start_idx, do_snapshots=True, do_prune=True):
+    open_last_row = 0
+    for i in range(total_appends):
+        first_row = start_idx + i * rows_per_append
+        open_last_row = start_idx + (i + 1) * rows_per_append
+        snap = 'snap_{}'.format(first_row)
+        library.append(symbol, bulk_data_ts[first_row:open_last_row],
+                       metadata={'snap': snap},
+                       prune_previous_version=do_prune)
+        if do_snapshots:
+            library.snapshot(snap)
+
+    return open_last_row
+
+
+def test_no_corruption_restore_append_overlapping(library, library_name):
+    large_ts = create_test_data(size=3000, cols=100,
+                                index=True, multiindex=False,
+                                random_data=True, random_ids=True)
+    rows_per_append = 100
+    with patch('pymongo.message.query', side_effect=_query(False, library_name)), \
+         patch('pymongo.server_description.ServerDescription.server_type', SERVER_TYPE.Mongos):
+
+        n_append(library, library_name, 18, rows_per_append, large_ts, 0)
+
+        # Corrupts all versions between the version that row "restore_from_row" was written,
+        restore_from_row = rows_per_append * 10
+        library.restore_version(symbol, 'snap_{}'.format(restore_from_row))
+        library.append(symbol, large_ts[restore_from_row:restore_from_row + 50])
+
+        # last_v = library._versions.find_one(sort=[('version', pymongo.DESCENDING)])
+        # analyze_symbol(library, symbol, 0, last_v['version'] + 1)
+
+        # Verify no versions have been corrupted
+        for v in library._versions.find(sort=[('version', pymongo.DESCENDING)]):
+            library.read(symbol, as_of=v['version'])
+
+
+def test_no_corruption_restore_writemeta_append(library, library_name):
+    large_ts = create_test_data(size=2000, cols=100,
+                                index=True, multiindex=False,
+                                random_data=True, random_ids=True)
+    rows_per_append = 100
+    with patch('pymongo.message.query', side_effect=_query(False, library_name)), \
+         patch('pymongo.server_description.ServerDescription.server_type', SERVER_TYPE.Mongos):
+
+        last_row = n_append(library, library_name, 9, rows_per_append, large_ts, 0)
+
+        library.write_metadata(symbol, metadata={'abc': 'xyz'})
+
+        n_append(library, library_name, 9, rows_per_append, large_ts, last_row)
+
+        library.write_metadata(symbol, metadata={'abc2': 'xyz2'})
+
+        # Corrupts all versions between the version that row "restore_from_row" was written,
+        restore_from_row = rows_per_append * 10
+        library.restore_version(symbol, 'snap_{}'.format(restore_from_row))
+
+        library.write_metadata(symbol, metadata={'abc3': 'xyz3'})
+
+        library.append(symbol, large_ts[restore_from_row:restore_from_row + 50])
+
+        library.write_metadata(symbol, metadata={'abc4': 'xyz4'})
+
+        # analyze_symbol(library, symbol, 0, last_v['version'] + 1)
+
+        # Verify no versions have been corrupted
+        for v in library._versions.find(sort=[('version', pymongo.DESCENDING)]):
+            library.read(symbol, as_of=v['version'])
+
+
+def test_no_corruption_restore_append_non_overlapping_tstamps(library, library_name):
+    large_ts = create_test_data(size=2000, cols=100,
+                                index=True, multiindex=False,
+                                random_data=True, random_ids=True)
+    with patch('pymongo.message.query', side_effect=_query(False, library_name)), \
+         patch('pymongo.server_description.ServerDescription.server_type', SERVER_TYPE.Mongos):
+
+        # Append with 50 small uncompressed segments (no new base yet)
+        last_row_b = n_append(library, library_name, 50, 25, large_ts, 0, False, True)
+
+        library.snapshot('snap_A')
+
+        # Append with 20 more small segments, causes once copy-rewrite with new base, and then some small appended segments
+        n_append(library, library_name, 15, 25, large_ts, last_row_b, True, True)
+
+        library.restore_version(symbol, as_of='snap_A')
+
+        last_row = n_append(library, library_name, 1, 40, large_ts, last_row_b, False, True)
+        library.snapshot('snap_B')
+
+        # Corrupts all versions
+        last_row = n_append(library, library_name, 1, 10, large_ts, last_row, False, True)
+        last_row = n_append(library, library_name, 8, 20, large_ts, last_row, False, True)
+        library.snapshot('snap_C')
+        # last_v = library._versions.find_one(sort=[('version', pymongo.DESCENDING)])
+        # analyze_symbol(library, symbol, 0, last_v['version'] + 1)
+        # Verify no versions have been corrupted
+        for v in library._versions.find(sort=[('version', pymongo.DESCENDING)]):
+            library.read(symbol, as_of=v['version'])
+
+
+def test_restore_append_overlapping_corrupts_old(library, library_name):
+    large_ts = create_test_data(size=2000, cols=100,
+                                index=True, multiindex=False,
+                                random_data=True, random_ids=True)
+    with patch('pymongo.message.query', side_effect=_query(False, library_name)), \
+         patch('pymongo.server_description.ServerDescription.server_type', SERVER_TYPE.Mongos), patch('arctic.store.version_store'):
+        library.write(symbol, large_ts[0:1000])
+        library.snapshot('snap_write_a')
+
+        library.append(symbol, large_ts[1000:1010])
+
+        library.restore_version(symbol, as_of='snap_write_a', prune_previous_version=True)
+        library.append(symbol, large_ts[1000:1009])
+
+        # from arctic.store._version_store_utils import analyze_symbol
+        # last_v = library._versions.find_one(sort=[('version', pymongo.DESCENDING)])
+        # analyze_symbol(library, symbol, 0, last_v['version'] + 1)
+
+        # Verify no versions have been corrupted
+        for v in library._versions.find(sort=[('version', pymongo.DESCENDING)]):
+            library.read(symbol, as_of=v['version'])
+
+
+def test_restore_append_overlapping_corrupts_last(library, library_name):
+    large_ts = create_test_data(size=2000, cols=100,
+                                index=True, multiindex=False,
+                                random_data=True, random_ids=True)
+    with patch('pymongo.message.query', side_effect=_query(False, library_name)), \
+         patch('pymongo.server_description.ServerDescription.server_type', SERVER_TYPE.Mongos), patch('arctic.store.version_store'):
+        library.write(symbol, large_ts[0:1000])
+        library.snapshot('snap_write_a')
+
+        library.append(symbol, large_ts[1000:1010])
+
+        library.restore_version(symbol, as_of='snap_write_a', prune_previous_version=True)
+        library.append(symbol, large_ts[1000:1012])
+
+        # from arctic.store._version_store_utils import analyze_symbol
+        # last_v = library._versions.find_one(sort=[('version', pymongo.DESCENDING)])
+        # analyze_symbol(library, symbol, 0, last_v['version'] + 1)
+
+        # Verify no versions have been corrupted
+        for v in library._versions.find(sort=[('version', pymongo.DESCENDING)]):
+            library.read(symbol, as_of=v['version'])
+
+
+# This is not necessary to fix, but the Exception thrown is quite confusing.
+@pytest.mark.skip(reason="Not critical as upsert=False is rarely used. A more specific handling/exception is required here.")
+def test_append_fail_after_delete_noupsert(library, library_name):
+    large_ts = create_test_data(size=2000, cols=100,
+                                index=True, multiindex=False,
+                                random_data=True, random_ids=True)
+    with patch('pymongo.message.query', side_effect=_query(False, library_name)), \
+         patch('pymongo.server_description.ServerDescription.server_type', SERVER_TYPE.Mongos), patch(
+        'arctic.store.version_store'):
+        library.write(symbol, large_ts[0:1000])  #v1
+        library.snapshot('snap_a')
+        library.append(symbol, large_ts[1000:1010])  #v2
+        library.snapshot('snap_b')
+        library.append(symbol, large_ts[1010:1020])  #v3
+        library.snapshot('snap_c')
+
+        library.append(symbol, large_ts[1030:1040])  #v4
+
+        library.delete(symbol) #v5
+
+        library.append(symbol, large_ts[1040:1050], upsert=False)  # v6
+        # from arctic.store._version_store_utils import analyze_symbol
+        # last_v = library._versions.find_one(sort=[('version', pymongo.DESCENDING)])
+        # analyze_symbol(library, symbol, 0, last_v['version'] + 1)
+
+        # Verify no versions have been corrupted
+        for v in library._versions.find(sort=[('version', pymongo.DESCENDING)]):
+            library.read(symbol, as_of=v['version'])
+
+
+def test_append_without_corrupt_check(library, library_name):
+    arctic.store._ndarray_store.set_corruption_check_on_append(False)
+    try:
+        with patch('pymongo.message.query', side_effect=_query(False, library_name)), \
+             patch('pymongo.server_description.ServerDescription.server_type', SERVER_TYPE.Mongos), patch('arctic.store.version_store'):
+            with pytest.raises(OperationFailure):
+                _corrupt_with_append_only(library, library_name)
+    finally:
+        arctic.store._ndarray_store.set_corruption_check_on_append(False)
+
+
+def test_append_with_corrupt_check(library, library_name):
+    arctic.store._ndarray_store.set_corruption_check_on_append(True)
+    try:
+        with patch('pymongo.message.query', side_effect=_query(False, library_name)), \
+             patch('pymongo.server_description.ServerDescription.server_type', SERVER_TYPE.Mongos), patch('arctic.store.version_store'):
+            _corrupt_with_append_only(library, library_name)
+    finally:
+        arctic.store._ndarray_store.set_corruption_check_on_append(False)
+
+
+def _corrupt_with_append_only(library, library_name):
+    def do_fail(version):
+        raise Exception('test')
+
+    large_ts = create_test_data(size=2000, cols=100,
+                                index=True, multiindex=False,
+                                random_data=True, random_ids=True)
+    library.write(symbol, large_ts[0:1000])  # v1
+    library.snapshot('snap_write_a')
+    library.append(symbol, large_ts[1000:1010])  # v2
+    library.snapshot('snap_write_b')
+
+    # Here we simulate a scenario where an append succeeds to insert the data segments,
+    # but fails to insert the version document (i.e. Mongo error occurred)
+    orig_insert_version = library._insert_version
+    library._insert_version = do_fail
+    try:
+        library.append(symbol, large_ts[1010:1020])  # v3
+    except:
+        pass
+    library._insert_version = orig_insert_version
+
+    library.write_metadata(symbol, {'hello': 'there'})  # , prune_previous_version=False)
+
+    # Appending subsequently overlapping and non-SHA-matching data cause data corruption
+    library.append(symbol, large_ts[1018:1030])  # , prune_previous_version=False)
+
+    # from arctic.store._version_store_utils import analyze_symbol
+    # last_v = library._versions.find_one(sort=[('version', pymongo.DESCENDING)])
+    # analyze_symbol(library, symbol, 0, last_v['version'] + 1)
+
+    # Verify no versions have been corrupted
+    for v in library._versions.find(sort=[('version', pymongo.DESCENDING)]):
+        library.read(symbol, as_of=v['version'])
+
+
+def test_snapshot_list_versions_after_delete(library, library_name):
+    with patch('pymongo.message.query', side_effect=_query(False, library_name)), \
+         patch('pymongo.server_description.ServerDescription.server_type', SERVER_TYPE.Mongos), patch(
+        'arctic.store.version_store'):
+        library.write("symA", 'data data')
+        library.write("symB", 'data data')
+        library.write("symC", 'data data')
+        library.snapshot('snapA')
+
+        library.delete('symC')
+
+        assert {v['symbol'] for v in library.list_versions(snapshot='snapA')} == {'symA', 'symB', 'symC'}